Test all frameworks (#266)
* Test all frameworks

* Move keras tests to tf1/tf2

* Single keras directory

* Stubs for tests

* KerasClassifier is only valid for TF1
