+++ .github/workflows/ci_test.yml
@@ -42,8 +42,39 @@ jobs:
         run: |
           pip install -r requirements.txt
           version=$(python -m armory --version)
-          bash docker/build-dev-minimal.sh
+          bash docker/build-dev-minimal.sh tf1
           docker run -w /armory_dev twosixarmory/tf1:${version} pytest -s ./tests/test_docker
+          docker run -w /armory_dev twosixarmory/tf1:${version} pytest -s ./tests/test_tf1
+  pytest-tf2:
+    name: PyTest TF2
+    runs-on: ubuntu-18.04
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/setup-python@v1
+        with:
+          python-version: '3.6'
+      - name: Run armory TF2 docker tests
+        run: |
+          pip install -r requirements.txt
+          version=$(python -m armory --version)
+          bash docker/build-dev-minimal.sh tf2
+          docker run -w /armory_dev twosixarmory/tf2:${version} pytest -s ./tests/test_docker
+          docker run -w /armory_dev twosixarmory/tf2:${version} pytest -s ./tests/test_tf2
+  pytest-pytorch:
+    name: PyTest PyTorch
+    runs-on: ubuntu-18.04
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/setup-python@v1
+        with:
+          python-version: '3.6'
+      - name: Run armory PyTorch docker tests
+        run: |
+          pip install -r requirements.txt
+          version=$(python -m armory --version)
+          bash docker/build-dev-minimal.sh pytorch
+          docker run -w /armory_dev twosixarmory/pytorch:${version} pytest -s ./tests/test_docker
+          docker run -w /armory_dev twosixarmory/pytorch:${version} pytest -s ./tests/test_pytorch
   flake8-test:
     name: Flake8
     runs-on: ubuntu-18.04
