+++ tests/test_pytorch/test_pytorch_models.py
@@ -1,0 +1,6 @@
+import unittest
+
+
+class PyTorchModelsTest(unittest.TestCase):
+    def test_pytorch_mnist(self):
+        pass
