+++ tests/test_tf1/test_tf1_models.py
@@ -1,0 +1,6 @@
+import unittest
+
+
+class TF1ModelsTest(unittest.TestCase):
+    def test_tf1_mnist(self):
+        pass
