+++ tests/test_tf2/test_tf2_models.py
@@ -1,0 +1,6 @@
+import unittest
+
+
+class TF2ModelsTest(unittest.TestCase):
+    def test_tf2_mnist(self):
+        pass
